Implemented first part of sub-service-requests; tests passing.
Several categories of changes:

The 'optional' attribute on line_items and service associations was
being set to a string ("true" or "false") instead of a boolean, so I
changed that wherever I found relevant code.

The associated services were being handled in a somewhat messy fashion:
there was an AssociatedServiceList collection that would query the
server and bring back a list of service-association maps which weren't
really Service models. The CatalogView would then individually query the
server for the actual services. I first of all changed the names in the
routes/models/etc from AssociatedService to ServiceAssociation, to
reflect the fact that the object is not a Service, and so also changed
the model on the collection to a vanilla Backbone.Model. Furthermore I
realized that at least in the context of the CatalogView, if we assume
that associated services are always in the same core/program (which I
think we can), then we should be able to find the service simply by
querying the other service's collection; so I changed the code to do
that.

Actually implemented the sub-service-request functionality, in the
ServiceRequest model on an event handler for LineItems#add. I created
a SubServiceRequest model to be used as a pseudo-model the same way
that LineItems are. It's worth looking into whether this pattern could
be refactored out into a model extension or something. I think we use
it in a third place somewhere as well but can't remember.

One of the unrelated tests that ended up failing after my implementation
was the one-time-fees test in the CatalogViewSpec.  Upon examining the
test, I realized that it was almost entirely a test of the model,
particularly considering how deeply it was reaching into the view and
pulling things out. Rather than fix the test, which could have been a
bit hairy, I decided to change it to a model test. This necessitated
moving some logic from the CatalogView into the ServiceRequest. I think
some more refactoring along those lines should be done, and it could
help clean up the CatalogView considerably, as well as make it easier
for us to test the business logic.