Rewrite most of the associated service code
For maximum flexibility we want to have the associated services
of each service cached just like all the other catalog data.
My solution was to assume that every service comes from the
server with an 'associations' key, the value of which is an array
of objects like {"service_id": "abc", "optional": true}. This
will present some challenges on the server side but I think by
adding a route to ObisEntity it can be done efficiently. The
major client-side changes were fetch hooks for when services
are loaded, and related factory/mocking/spec stuff.

While I was at it I went ahead and refactored the CartView,
since a lot of the association code was there and in the catalog
view. I moved all the association code to the models, and
moved all the functionality that was in the global "CartBuilder"
to the CartView itself, which wasn't doing much otherwise. The
new structure should be a lot more modular, backboney, and simple
to test.

Corresponding server side changes are pending, so this commit
isn't a functional snapshot.